Complain if PRs don't reference regions correctly (#3)
We want to be able to rely on regions for filtering, e.g. when showing
things in dashboards.

Ideally we would have a source of truth of which region each trainee is
in, but we don't, so in the mean time check that the title claims to be
in some region.

Hard-code the known regions for now - ideally an API will provide these
in the future.