Add potentially ignorable permissions errors
There are times where if we can get extra information we want to show
it, but if not we want to show what information we have. For instance,
we currently have a staff view and a non-staff view of the "reviewers"
dashboard. This construction will allow us to just try to get the
staff-only information, and if it can be fetched, show it, and if not,
just not show it. We can collapse two views into one. The down side here
is if someone _should_ have access to the staff view, they may not
notice that they're missing some extra data.