Fix attendance sheet filtering (#41)
This bug was introduced in aa0d5787598ab8191f6429b6858f3b62dd2a6f9f - it
used to continue out of an outer loop to the next sheet, but now the
continue just skips to the next row.

This function now handles just one sheet, so if a sheet is not a
register sheet, just return immediately.