Expire sessions after 1 hour
It looks like something changed with either GitHub or browsers such that
GitHub is frequently giving 401s. I'm not sure whether that's because
they're expiring tokens more readily, or sessions are lasting longer
than they used to.

Ideally Octocrab would notice 401s and force a re-auth, but it doesn't
have an obvious hook for that. Instead, just expire sessions after an
hour so that we will stop seeing existing tokens, and will automatically
redirect people to fetch new tokens.